Linting and reorganization
