+++ include/Action.hpp
@@ -1,0 +1,10 @@
+// Copyright 2019 Jack Khuu, Justin Johnson, Lukas Joswiak
+
+#ifndef INCLUDE_ACTION_HPP_
+#define INCLUDE_ACTION_HPP_
+
+namespace Action {
+enum Action { PUT, APPEND, GET };
+}  // namespace Action
+
+#endif  // INCLUDE_ACTION_HPP_
