--- src/util/KVStore/AMOStore.h
@@ -1,29 +1,0 @@
-#ifndef _AMOSTORE_H_
-#define _AMOSTORE_H_
-
-#include <map>
-#include <string>
-
-#include "./AMOCommand.h"
-#include "./AMOResponse.h"
-#include "./KVStore.h"
-
-namespace AMOStore {
-using namespace std;
-
-class AMOStore {
-  public:
-    AMOStore();
-	  
-    AMOResponse::AMOResponse Execute(const AMOCommand::AMOCommand &command);
-
-  private:
-    KVStore::KVStore kvStore_;
-    map<AMOCommand::AMOCommand, AMOResponse::AMOResponse> prev_;
-
-    bool AlreadyExecuted(const AMOCommand::AMOCommand &command) const;
-};
-
-} //namespace AMOStore
-
-#endif // _AMOSTORE_H_
