--- src/util/KVStore/Action.h
@@ -1,8 +1,0 @@
-#ifndef _ACTION_H_
-#define _ACTION_H_
-
-namespace Action {
-enum Action { PUT, APPEND, GET };
-} // namespace Action
-
-#endif // _ACTION_H_
