+++ src/util/KVStore/brainstorm.cpp
@@ -1,150 +1,157 @@
+// Copyright 2019 Jack Khuu, Justin Johnson, Lukas Joswiak
+
+# include <stdlib.h>
+
 # include <map>
 # include <string>
-# include <stdlib.h>
 # include <iostream>
 
-using namespace std;
+using std::map;
+using std::string;
+using std::cout;
+using std::endl;
 
 enum Action { PUT, APPEND, GET };
 
 // Supports an integer Key and a String value
 class KVStore {
-  private:
-    map<int, string> store_;
+ private:
+  map<int, string> store_;
 
 
-  public:
-    string insert(const int &key, const string &value) {
-      if (store_.find(key) != store_.end()) {
-        exit(EXIT_FAILURE);
-      }
-      store_[key] = value;
-      
-      return store_[key];
+ public:
+  string insert(const int &key, const string &value) {
+    if (store_.find(key) != store_.end()) {
+      exit(EXIT_FAILURE);
     }
+    store_[key] = value;
 
-    string append(const int &key, const string &value) {
-      if (store_.find(key) == store_.end()) {
-        exit(EXIT_FAILURE);
-      }
-      store_[key] = store_[key] + value;  
-      return store_[key];
-    }
+    return store_[key];
+  }
 
-    string get(const int &key) {
-      return store_[key];
+  string append(const int &key, const string &value) {
+    if (store_.find(key) == store_.end()) {
+      exit(EXIT_FAILURE);
     }
+    store_[key] = store_[key] + value;
+    return store_[key];
+  }
+
+  string get(const int &key) {
+    return store_[key];
+  }
 };
 
 
 class AMOCommand {
-  private:
-    int seqNum_;
-    int key_;
-    string value_;
-    Action action_;
-    
+ private:
+  int seqNum_;
+  int key_;
+  string value_;
+  Action action_;
 
-  public:
-    AMOCommand() {}
 
-    AMOCommand(const int &seqNum, const int &key, const string &value, const Action &action) {
-      seqNum_ = seqNum;
-      key_ = key;
-      value_ = value;
-      action_ = action;
-    }
+ public:
+  AMOCommand() {}
 
-    bool operator< (const AMOCommand &r) const {
-      return this->seqNum_ < r.seqNum_;
-    }
+  AMOCommand(const int &seqNum, const int &key, const string &value,
+                  const Action &action) {
+    seqNum_ = seqNum;
+    key_ = key;
+    value_ = value;
+    action_ = action;
+  }
 
-    int getSeqNum() const {
-      return seqNum_;
-    }
+  bool operator< (const AMOCommand &r) const {
+    return this->seqNum_ < r.seqNum_;
+  }
 
-    int getKey() const {
-      return key_;
-    }
+  int getSeqNum() const {
+    return seqNum_;
+  }
 
-    string getValue() const {
-      return value_;
-    }
+  int getKey() const {
+    return key_;
+  }
 
-    Action getAction() const {
-      return action_;
-    }
+  string getValue() const {
+    return value_;
+  }
+
+  Action getAction() const {
+    return action_;
+  }
 };
 
 
 class AMOResponse {
-  private:
-    AMOCommand command_;
-    string value_; 
-      
-  public:
-    AMOResponse() {}
+ private:
+  AMOCommand command_;
+  string value_;
 
-    AMOResponse(const AMOCommand &command, const string &value) {
-      command_ = command;
-      value_ = value; 
-    }
+ public:
+  AMOResponse() {}
 
-    bool operator< (const AMOResponse &r) const {
-      return command_ < r.command_;
-    }
+  AMOResponse(const AMOCommand &command, const string &value) {
+    command_ = command;
+    value_ = value;
+  }
 
-    AMOCommand getCommand() const {
-      return command_;
-    }
+  bool operator< (const AMOResponse &r) const {
+    return command_ < r.command_;
+  }
 
-    string getValue() const {
-      return value_;
-    }
+  AMOCommand getCommand() const {
+    return command_;
+  }
+
+  string getValue() const {
+    return value_;
+  }
 };
 
 
 class AMOStore {
-  private:
-    KVStore kvStore_;
-    map<AMOCommand, AMOResponse> prev_;
+ private:
+  KVStore kvStore_;
+  map<AMOCommand, AMOResponse> prev_;
 
-    bool alreadyExecuted(const AMOCommand &command) const {
-      return prev_.find(command) != prev_.end();
-    }
+  bool alreadyExecuted(const AMOCommand &command) const {
+    return prev_.find(command) != prev_.end();
+  }
 
-  public:
-    AMOStore() {
-      kvStore_ = KVStore();
-    }
-	  
-    AMOResponse execute(const AMOCommand &command) {
-      string s;
-      if (!alreadyExecuted(command)) {
-	cout << "New" << endl;
-	switch(command.getAction()) {
-	  case PUT:
-	    cout << "PUT" << endl;
-            s = kvStore_.insert(command.getKey(), command.getValue());
-	    break;
-	  case APPEND:
-	    cout << "APPEND" << endl;
-            s = kvStore_.append(command.getKey(), command.getValue());
-	    break;
-	  case GET:
-	    cout << "GET" << endl;
-            s = kvStore_.get(command.getKey());
-	    break;
-	  default:  
-	    exit(EXIT_FAILURE);
-	}
+ public:
+  AMOStore() {
+    kvStore_ = KVStore();
+  }
 
-	prev_[command] = AMOResponse(command, s);
+  AMOResponse execute(const AMOCommand &command) {
+    string s;
+    if (!alreadyExecuted(command)) {
+      cout << "New" << endl;
+      switch (command.getAction()) {
+        case PUT:
+          cout << "PUT" << endl;
+          s = kvStore_.insert(command.getKey(), command.getValue());
+          break;
+        case APPEND:
+          cout << "APPEND" << endl;
+          s = kvStore_.append(command.getKey(), command.getValue());
+          break;
+        case GET:
+          cout << "GET" << endl;
+          s = kvStore_.get(command.getKey());
+          break;
+        default:
+          exit(EXIT_FAILURE);
       }
-      cout << "Returned from Execute: ";
-      cout << prev_[command].getValue() << endl;
-      return prev_[command];
+
+      prev_[command] = AMOResponse(command, s);
     }
+    cout << "Returned from Execute: ";
+    cout << prev_[command].getValue() << endl;
+    return prev_[command];
+  }
 };
 
 
@@ -180,7 +187,6 @@ int main() {
   app.execute(c1a);
   app.execute(c1g);
   app.execute(c1g2);
-  
 
   return 0;
 }
