+++ src/util/KVStore/example.cpp
@@ -1,16 +1,18 @@
+// Copyright 2019 Jack Khuu, Justin Johnson, Lukas Joswiak
+
 #include <iostream>
 
-#include "./AMOStore.h"
-#include "./AMOCommand.h"
-#include "./Action.h"
+#include "AMOStore.hpp"
+#include "AMOCommand.hpp"
+#include "Action.hpp"
 
-using namespace Action;
-using namespace std;
+using std::cout;
+using std::endl;
 
 int main() {
   AMOStore::AMOStore app = AMOStore::AMOStore();
 
-  AMOCommand::AMOCommand c1 = AMOCommand::AMOCommand(0, 10, "A", PUT);
+  AMOCommand::AMOCommand c1 = AMOCommand::AMOCommand(0, 10, "A", Action::PUT);
   app.Execute(c1);
   cout << "Done" << endl;
   return 0;
